Fix bug and add slide to uploaded photo
